requirements.txt
passlib
resend
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
alembic
psycopg2-binary
asyncpg
sqlalchemy[asyncio]
fastapi
uvicorn
python-dotenv
alembic
pymysql